Since features are in memory, use that instead of hitting the database.

This also fixes a bug where hidden features were still displayed
and fixes a bug where sibling features were still displayed.


empty dashboard - 3rd refresh - saves 38 sql queries and 60ms per page.

Rendering Before|time|queries|db time
---|---|---|---|---
`layouts/_page_header_navbar`|58.4ms|43 sql|8.0ms
Response|328ms

Rendering After|time|queries|db time
---|---|---|---|---
`layouts/_page_header_navbar`|10.2ms|3 sql|0.6ms
Response|285ms

/cc @jrafanie @Fryguy @dmetzger57 here you go
/cc @h-kataria thanks for help