@kbrock 2015-11-30
@Fryguy @chessbyte suggestions on making easier to merge? `feature_children` and `feature_details` had a bug that was not hiding hidden nodes. Also there is a performance fix here.