@kbrock 2015-11-30
http://localhost:3000/vm_infra/explorer 12vms production mode

rendering before|time|queries|db time
---|---|---|---
explorer       |233ms| 41|21.6ms
user_options   |  3ms|  1| 0.3ms
page_header_navbar         |  65ms | 41|7.4ms
x_taskbar      | 83ms | 27|6ms
adv_search_body| 3ms |  1|0.3ms

rendering after|time|queries|db time
---|---|---|---
explorer           | 190ms|38|19.8ms
user_options       |  3ms|  1| 0.3ms
page_header_navbar |  7ms|1|0.2ms
x_taskbar          | 12.9ms |0|0ms
adv_search_body    | 2.3ms | 1|0.4ms